backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# PDF processing
pdfplumber==0.10.3

# Embedding and ML models
sentence-transformers==2.6.1
torch>=1.13.0
transformers>=4.21.0

# LLM API
google-generativeai>=0.3.0

# Vector database
chromadb==0.4.18

# Data handling and validation
pydantic==2.5.0
numpy>=1.24.0

# Optional: LangChain (if you want to use it)
# langchain==0.1.0
# langchain-google-genai==0.0.6
# langchain-community==0.0.10
